Convert APIWorker to typescript and remove duplicate call to all APIs on app load
